% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmd_utils.R
\name{calc_date_breaks}
\alias{calc_date_breaks}
\title{calc_date_breaks}
\usage{
calc_date_breaks(first_date, last_date)
}
\arguments{
\item{first_date}{a date or string that as.Date can convert to date}

\item{last_date}{a date or string that as.Date can convert to date}
}
\value{
A character string
}
\description{
Helper function to help calculate the number of breaks needed an a date axis.
The output is a string in the format "N days" suitable as value for parameter
date_breaks in scale_x_date() and friends.
}
\examples{
calc_date_breaks("2022-06-01", "2022-06-18")
}
