% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fix_fangst_missing_fangstdattid}
\alias{fix_fangst_missing_fangstdattid}
\title{Fix missing datetime on fangster}
\usage{
fix_fangst_missing_fangstdattid(f, r)
}
\arguments{
\item{f}{data.frame with fangster as returned by read_fangst_clean()}

\item{r}{data.frame with resor as returned by read_resa_clean()}
}
\value{
a data.frame where catches with missing FANGSTDATTID have their FANGSTDATTID set
to RESEDATUM "12:00:00"
}
\description{
Fix missing datetime on fangster
}
\examples{
\dontrun{
fangst <- fix_fangst_missing_fangstdattid(fangst, resa)
}
}
