% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmd_utils.R
\name{integer_breaks}
\alias{integer_breaks}
\title{integer_breaks}
\usage{
integer_breaks(n = 5, ...)
}
\arguments{
\item{n}{Number of wanted breaks}

\item{...}{passed to scales::pretty_breaks}
}
\value{
a function
}
\description{
Put breaks only om integer values
}
