% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmd_utils.R
\name{minutes2HHMM}
\alias{minutes2HHMM}
\title{Minutes to HH:MM}
\usage{
minutes2HHMM(minutes)
}
\arguments{
\item{minutes}{}
}
\value{
character
}
\description{
Convert number of minutes to a character string in format HH:;;
}
\examples{
minutes2HHMM(c(61, 150))
}
