% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APEX_data.R
\name{read_anvlista}
\alias{read_anvlista}
\title{Read a file with registered sporeg users}
\usage{
read_anvlista()
}
\value{
Return a tibble
}
\description{
The location of the file read is predefined with but can be changed with
the function ?APEX_options().
The function uses file extension to select how the file should be read.
Known extensions are .csv and .xlsx.
#'
}
\examples{
\dontrun{
  users <- read_anvlista()
}
}
