% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_fangst_clean}
\alias{read_fangst_clean}
\title{Read a CSV-file with APEX export of fångster from Spöreg database and do some cleanup}
\usage{
read_fangst_clean(file_name = "Spöreg Fångst.csv")
}
\arguments{
\item{file_name}{character Default "Spöreg Fångst.csv"}
}
\value{
Return a tibble with all catches in file_name
}
\description{
Read a CSV-file with APEX export of fångster from Spöreg database and do some cleanup
}
\examples{
\dontrun{
fangst <- read_fangst_clean()
}
}
