% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_resa_clean}
\alias{read_resa_clean}
\title{Read a CSV-file with APEX export of resor from Spöreg database and do some cleanup}
\usage{
read_resa_clean(file_name = "Spöreg Resa.csv")
}
\arguments{
\item{file_name}{character Default "Spöreg Resa.csv"}
}
\value{
Return a tibble with all trips in file_name
}
\description{
Read a CSV-file with APEX export of resor from Spöreg database and do some cleanup
}
\examples{
\dontrun{
  resor <- read_resa_clean()
}
}
