% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_utils.R
\name{sporeg_bbox}
\alias{sporeg_bbox}
\title{sporeg_bbox}
\usage{
sporeg_bbox(x, lat = "POSITIONN", lon = "POSITIONE")
}
\arguments{
\item{x}{Data frame containing coordinates}

\item{lat}{character Column name or number containing latitude. Default "POSITIONN"}

\item{lon}{character Column name or number containing longitude. Default "POSITIONE"}
}
\value{
Returns a bounding box as a vector of four named numbers. The numbers are named:
"left", "bottom", "right", and "top". This is the format used by ggmap::get_map()
}
\description{
Return bounding box for data frame. The data must contain two columns that
contains coordinates
}
