% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_utils.R
\name{sporeg_points}
\alias{sporeg_points}
\title{sporeg_points}
\usage{
sporeg_points(x, lat = "POSITIONN", lon = "POSITIONE")
}
\arguments{
\item{x}{Data frame containing coordinates}

\item{lat}{character Column name or number containing latitude. Default "POSITIONN"}

\item{lon}{character Column name or number containing longitude. Default "POSITIONE"}
}
\value{
Returns a sf multipoint object.
}
\description{
Create an sf-multipoint from a data frame. The data must contain two columns that
contains coordinates. Rows with NA in coordinates will be dropped silently.
}
