% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{sporegApp}
\alias{sporegApp}
\title{sporegApp}
\usage{
sporegApp()
}
\description{
Start the sporeg shiny app.
}
\details{
Browse data from the Spöreg app in several different ways. Input data
is read from a predefined location (and filenames) that can be changed with
the \code{\link[sporegr]{APEX_options}} function.
}
\examples{
\dontrun{
library(sporeg)
# APEX_options(root_folder = "C:/your/folder") # if you have private copies
sporegApp()
}
}
