% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sporegr-package.R
\docType{package}
\name{sporegr-package}
\alias{sporegr}
\alias{sporegr-package}
\title{sporegr: Work With Data From the Spöreg mobile app}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Anders Kagervall \email{anders.kagervall@slu.se} (\href{https://orcid.org/0000-0003-4790-2696}{ORCID})

}
\keyword{internal}
