% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_color_sporeg.R
\name{sporegr_scales}
\alias{sporegr_scales}
\alias{scale_color_sporeg}
\alias{scale_fill_sporeg}
\alias{scale_color_sporeg_art}
\title{Create discrete scales for sporegr}
\usage{
scale_color_sporeg(...)

scale_fill_sporeg(...)

scale_color_sporeg_art(...)
}
\arguments{
\item{...}{}
}
\value{
a ggplot2 color scale
}
\description{
These functions set manual ggplot color scales conditional on names in Spöreg.
}
\details{
Replaces the use of e.g. "scale_colour_manual()" and "scale_fill_manual().
pal <- c("#56B4E9", "#009E73", "#F0E442", "#0072B2", "#E69F00", "#D55E00")

This ensures that each FANGOMR gets predesignated colors according to the standard in Fiskbarometern.
Could potentially also be used for ARTBEST/MALART
}
