% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APEX_data.R
\name{APEX_options}
\alias{APEX_options}
\title{Set or get options for reading APEX Spöreg data}
\usage{
APEX_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for reading APEX Spöreg data
}
\section{Supported options}{

Usually you only need to change \code{root_folder} but it is possible to also change file names
if needed. Supported keys:
\itemize{
\item{\code{root_folder} Datafiles root (Default: //storage-dh.slu.se/restricted$/Stora sjoarna/Data/Fritidsfiskedata/Spöreg)}
\item{\code{resa} File with trip occasions (Default: Spöreg Resa.xlsx)}
\item{\code{fangst} File with catch/recatch data (Default: Spöreg Resa.xlsx)}
\item{\code{ovrighandelse} File with misc events (Default: Spöreg Övrighändelse.xlsx)}
\item{\code{anvlista} File with username, full name and organisation (Default: anvlista.xlsx)}
}
}

\examples{
# Show default
APEX_options()

APEX_options(root_folder = "C:/myfolder") # Change folder where all files are located
APEX_options()$root_folder # Show the new value
settings::reset(APEX_options) # Reset to defaults

}
