% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{read_fangst_clean}
\alias{read_fangst_clean}
\title{Read a file with APEX export of fångster from Spöreg database and do some cleanup}
\usage{
read_fangst_clean(file_name = "Spöreg Fångst.csv")
}
\arguments{
\item{file_name}{character Default "Spöreg Fångst.csv"}
}
\value{
Return a tibble
}
\description{
The function uses file extension to select how the file should be read.
Known extensions are .csv and .xlsx.
}
\examples{
\dontrun{
fangst <- read_fangst_clean()
}
}
