% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APEX_data.R
\name{read_fangst_clean}
\alias{read_fangst_clean}
\title{Read a file with APEX export of fångster from Spöreg database and do some cleanup}
\usage{
read_fangst_clean()
}
\value{
Return a tibble
}
\description{
The function uses file extension to select how the file should be read.
}
\examples{
\dontrun{
fangst <- read_fangst_clean()
}
}
