% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APEX_data.R
\name{read_ovrighandelse_clean}
\alias{read_ovrighandelse_clean}
\title{Read a file with APEX export of Övrighändelse from Spöreg database and do some cleanup}
\usage{
read_ovrighandelse_clean()
}
\value{
Return a tibble
}
\description{
The function uses file extension to select how the file should be read.
Known extensions are .csv and .xlsx.
}
\examples{
\dontrun{
ovr <- read_ovrighandelse_clean()
}
}
