% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APEX_data.R
\name{read_resa_clean}
\alias{read_resa_clean}
\title{Read a file with APEX export of resor from Spöreg database and do some cleanup}
\usage{
read_resa_clean()
}
\value{
Return a tibble
}
\description{
The location of the file read is predefined with but can be changed with
the function \code{\link[=APEX_options]{APEX_options()}}.
The function uses file extension to select how the file should be read.
}
\examples{
\dontrun{
  resor <- read_resa_clean()
}
}
