% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sporegr-package.R
\docType{package}
\name{sporegr-package}
\alias{sporegr}
\alias{sporegr-package}
\title{sporegr: tools and templates to work with and create reports from Spöreg data}
\description{
This package will help you Work with data from the Spöreg mobile app. The package contains
functions to read (and do some cleanup) of data files exported from predefined queries in APEX.
}
\section{Data Import and Export}{


Functions to import data from csv or xlsx files exported from the APEX
service (\url{https://fd2.slu.se/ords/r/aqua/store100107/home}). Some cleanup
and renaming is done to make the dataset more R friendly.

\itemize{
\item \code{\link{read_resa_clean}}:  import data from with trip info (\emph{resor})
\item \code{\link{read_fangst_clean}}:  import data from with catches (\emph{fångst})
\item \code{\link{read_fangst_ovrighandelse}}:  import data from with misc data
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SLU-Aqua-large-lakes/sporegr}
  \item \url{https://slu-aqua-large-lakes.github.io/sporegr/}
  \item Report bugs at \url{https://github.com/SLU-Aqua-large-lakes/sporegr/issues}
}

}
\author{
\strong{Maintainer}: Göran Sundblad \email{goran.sundblad@slu.se} (\href{https://orcid.org/0000-0001-8970-9996}{ORCID})

Authors:
\itemize{
  \item Anders Kagervall \email{anders.kagervall@gmail.com} (\href{https://orcid.org/0000-0003-4790-2696}{ORCID})
}

}
\keyword{internal}
